requirements.txt
aiogram==2.9.2
python-dotenv==0.14.0
requests==2.24.0